third_party/blink/web_tests/flag-specific/disable-layout-ng/accessibility/aria-owns-sends-notification-expected.txt

This is a testharness.js-based test.
PASS A children changed notification is fired when an aria-owned child gets added to the tree.
FAIL A children changed notification is fired when an aria-owned attribute is added to an element that causes it to reparent another element to become its child. assert_equals: expected "ChildrenChanged" but got "MarkDirty"
PASS A children changed notification is fired on the old parent when one of its children gets reparented to another element due to aria-owns.
Harness: the test ran to completion.
